+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 #### Version 0.9.4 - ???
 * Point github icon to https://github.com/wabbajack-tools/wabbajack
+* Add game registry entry for Skyrim VR
+* Modlists are now .zip files. 
+* Modlists now end with `.modlist_v1` to enable better version control
 
 #### Version 0.9.3 - 9/30/2019
 * Add WABBAJACK_NOMATCH_INCLUDE works like WABBAJACK_INCLUDE but only includes files that are found to be missing at the end of compilation